fix(e2e): close the race windows in the new live tests
The fail-closed test now waits for ResolvedRefs=False/UnsupportedProtocol
BEFORE treating a 200 as fatal -- the status write happens after the
config push, so an observed condition guarantees the fail-closed config
reached the proxy and a 200 can no longer be the stale-cache window of
an early reconcile. The gRPC TLS test moves its serving-pod check
inside the poll: an Echo answered by a foreign pod (a lingering
cleartext route on the same method) retries instead of failing. The
shared applyObject helper moves to setup_test.go with its siblings.
